pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "darkfield-defects"
version = "0.1.0"
description = "暗场显微镜离焦微结构镜片缺陷检测、分类与磨损评估系统"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "opencv-contrib-python>=4.8",
    "scipy>=1.10",
    "scikit-image>=0.21",
    "matplotlib>=3.7",
    "pillow>=10.0",
    "typer>=0.9",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.4", "mypy>=1.8"]
ml = ["torch>=2.0", "torchvision>=0.15", "segmentation-models-pytorch>=0.3", "ultralytics>=8.0"]
viz = ["gradio>=4.0", "ultralytics>=8.0"]

[project.scripts]
darkfield-defects = "darkfield_defects.cli.app:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
